+++ packages/moonstone/CHANGELOG.md
@@ -23,6 +23,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/ContextualPopupDecorator` to set accessibility-related props onto the container node rather than the popup node
 - `moonstone/Spinner` to position the text content under the spinner, rather than to the right side
 - `moonstone/VideoPlayer` to include hour when announcing the time while scrubbing
+- `moonstone/VideoPlayer` to handle play, pause, stop, fast forward and rewind on remote controller.
 
 ### Fixed